Move messages into standalone folder
