+++ src/lib/modules/basic-messages/index.ts
@@ -1,2 +1,2 @@
-export { BasicMessage } from './BasicMessage';
+export { BasicMessage } from './messages/BasicMessage';
 export { BasicMessageService, BasicMessageEventType, BasicMessageReceivedEvent } from './BasicMessageService';
