+++ src/lib/modules/basic-messages/messages/index.ts
@@ -1,0 +1,2 @@
+export * from './BasicMessage';
+export * from './BasicMessageMessageType';
