+++ src/lib/modules/routing/ProviderRoutingService.ts
@@ -3,8 +3,7 @@ import { OutboundMessage } from '../../types';
 import { createOutboundMessage } from '../../agent/helpers';
 import { InboundMessageContext } from '../../agent/models/InboundMessageContext';
 import { ConnectionRecord } from '../../storage/ConnectionRecord';
-import { KeylistUpdateMessage, KeylistUpdateAction } from './KeylistUpdateMessage';
-import { ForwardMessage } from './ForwardMessage';
+import { KeylistUpdateMessage, KeylistUpdateAction, ForwardMessage } from './messages';
 
 export interface RoutingTable {
   [recipientKey: string]: ConnectionRecord | undefined;
