+++ src/lib/modules/routing/handlers/ForwardHandler.ts
@@ -1,6 +1,6 @@
 import { Handler, HandlerInboundMessage } from '../../../handlers/Handler';
 import { ProviderRoutingService } from '../ProviderRoutingService';
-import { ForwardMessage } from '../ForwardMessage';
+import { ForwardMessage } from '../messages';
 
 export class ForwardHandler implements Handler {
   private routingService: ProviderRoutingService;
