+++ src/lib/modules/routing/messages/KeylistUpdateMessage.ts
@@ -2,8 +2,8 @@ import type { Verkey } from 'indy-sdk';
 import { Equals, IsArray, ValidateNested, IsString, IsEnum } from 'class-validator';
 import { Type } from 'class-transformer';
 
-import { AgentMessage } from '../../agent/AgentMessage';
-import { MessageType } from './messages';
+import { AgentMessage } from '../../../agent/AgentMessage';
+import { RoutingMessageType as MessageType } from './RoutingMessageType';
 
 export interface KeylistUpdateMessageOptions {
   id?: string;
