+++ src/lib/modules/routing/messages/RoutingMessageType.ts
@@ -1,4 +1,4 @@
-export enum MessageType {
+export enum RoutingMessageType {
   // TODO: add other messages from mediator coordination protocol
   KeylistUpdate = 'https://didcomm.org/coordinatemediation/1.0/keylist_update',
   BatchPickup = 'https://didcomm.org/messagepickup/1.0/batch-pickup',
